Add files via upload
When you clicked on OFF button , the light will off and there will be displayed a cat eyes. When you clicked on ON button ,the light will ON and entire will be seen.I had done this project with the help of HTML ,CSS , JavaScript. 
HERE IT IS WEBSITE LINK ,YOU CAN TRY IT IF YOU WANT : chandug7.ccbp.tech